NuGet package management.
https://www.nuget.org/packages/CreateNewNuGetPackageFromProjectAfterEachBuild/
won't work on account of this bug in NuGet: https://nuget.codeplex.com/workitem/4013
which won't be fixed until NuGet 3.0 is released.

However, it will be better to save the .nuspec file in source control and
use the windows nuget gui tool to produce the .nupkg file from that as an
untracked file on demand.